requirements.txt
flask
nltk
python-docx
gunicorn
reportlab  # New for PDF export
